+++ res/robots/movemaster/__init__.py
@@ -1,0 +1,1 @@
+print("[INFO] Importando librerias desde movemaster")
